+++ src/aac_metrics/__init__.py
@@ -6,13 +6,13 @@
 
 __name__ = "aac-metrics"
 __author__ = "Etienne Labbé (Labbeti)"
+__author_email__ = "labbeti.pub@gmail.com"
 __license__ = "MIT"
 __maintainer__ = "Etienne Labbé (Labbeti)"
 __status__ = "Development"
-__version__ = "0.1.1"
+__version__ = "0.1.2"
 
 
-from .functional.evaluate import aac_evaluate
 from .classes.coco_bleu import CocoBLEU
 from .classes.coco_cider_d import CocoCIDErD
 from .classes.coco_meteor import CocoMETEOR
@@ -20,3 +20,4 @@ from .classes.coco_rouge_l import CocoRougeL
 from .classes.coco_spice import CocoSPICE
 from .classes.evaluate import AACEvaluate
 from .classes.spider import SPIDEr
+from .functional.evaluate import aac_evaluate