+++ src/aac_metrics/classes/coco_cider_d.py
@@ -33,12 +33,14 @@ class CocoCIDErD(Metric):
         n: int = 4,
         sigma: float = 6.0,
         tokenizer: Callable[[str], list[str]] = str.split,
+        return_tfidf: bool = False,
     ) -> None:
         super().__init__()
         self._return_all_scores = return_all_scores
         self._n = n
         self._sigma = sigma
         self._tokenizer = tokenizer
+        self._return_tfidf = return_tfidf
 
         self._cooked_cands = []
         self._cooked_mrefs = []
@@ -50,6 +52,7 @@ class CocoCIDErD(Metric):
             self._return_all_scores,
             self._n,
             self._sigma,
+            self._return_tfidf,
         )
 
     def reset(self) -> None: