+++ src/aac_metrics/functional/common.py
@@ -1,27 +1,21 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-import subprocess
-
-from pathlib import Path
-from subprocess import CalledProcessError
-from typing import Any, Union
+from typing import Any
 
 
-def _check_java_path(java_path: Union[str, Path]) -> bool:
-    """Returns True if the java path is valid."""
-    if not isinstance(java_path, (str, Path)):
-        return False
+def is_mono_sents(sents: Any) -> bool:
+    """Returns True if input is list[str]."""
+    return isinstance(sents, list) and all(isinstance(sent, str) for sent in sents)
 
-    try:
-        exitcode = subprocess.check_call(
-            [java_path, "--version"],
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
-        )
-    except (CalledProcessError, PermissionError, FileNotFoundError):
-        exitcode = 1
-    return exitcode == 0
+
+def is_mult_sents(mult_sents: Any) -> bool:
+    """Returns True if input is list[list[str]]."""
+    return (
+        isinstance(mult_sents, list)
+        and all(isinstance(sents, list) for sents in mult_sents)
+        and all(isinstance(sent, str) for sents in mult_sents for sent in sents)
+    )
 
 
 def _check_input(
@@ -29,15 +23,10 @@ def _check_input(
     mult_references: Any,
 ) -> None:
     """Returns True candidates and mult_references have a valid type and size."""
-    cands_is_list = isinstance(candidates, list)
-    cand_is_str = all(isinstance(cand, str) for cand in candidates)
-    if not all((cands_is_list, cand_is_str)):
+    if not is_mono_sents(candidates):
         raise ValueError("Invalid candidates type. (expected list[str])")
 
-    mrefs_is_list = isinstance(mult_references, list)
-    refs_is_list = all(isinstance(refs, list) for refs in mult_references)
-    ref_is_str = all(isinstance(ref, str) for refs in mult_references for ref in refs)
-    if not all((mrefs_is_list, refs_is_list, ref_is_str)):
+    if not is_mult_sents(mult_references):
         raise ValueError("Invalid mult_references type. (expected list[list[str]])")
 
     same_len = len(candidates) == len(mult_references)