+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.2] 2022-10-31
+### Added
+- All candidates scores option `return_all_cands_scores` for SPIDEr-max.
+- Functions `is_mono_sents` and `is_mult_sents` to detect `list[str]` sentences and `list[list[str]]` multiples sentences.
+- Functions `flat_list` and `unflat_list` to flat multiples sentences to sentences.
+
+### Changed
+- Update default value used for `return_all_scores` in cider and rouge functions.
+- Update internal metric factory with functions instead of classes to avoid cyclic dependency.
+
+### Fixed
+- Fix SPIDEr-max local scores output shape.
+
 ## [0.1.1] 2022-09-30
 ### Added
 - Documentation for metric functions and classes.